Added failsafe for integration testing
