+++ mobtravelapp/src/main/java/com/mobiquity/mobtravelapp/model/travel/DestinationStub.java
@@ -1,19 +1,16 @@
-package com.mobiquity.mobtravelapp.model.travelModel;
+package com.mobiquity.mobtravelapp.model.travel;
 
+import com.mobiquity.mobtravelapp.model.weather.Weather;
 import lombok.Builder;
 import lombok.Getter;
 
 @Builder
 @Getter
-public class StopStub {
+public class DestinationStub {
     private final String plannedArrivalDateTime;
     private final String actualArrivalDateTime;
-    private final String plannedDepartureDateTime;
-    private final String actualDepartureDateTime;
     private final String plannedArrivalTrack;
     private final String actualArrivalTrack;
     private final Station station;
-
-
-
+    private final Weather weather;
 }
