+++ mobtravelapp/src/main/java/com/mobiquity/mobtravelapp/model/travel/RouteModel.java
@@ -1,4 +1,4 @@
-package com.mobiquity.mobtravelapp.model.travelModel;
+package com.mobiquity.mobtravelapp.model.travel;
 
 import lombok.*;
