+++ mobtravelapp/src/main/java/com/mobiquity/mobtravelapp/service/TravelService.java
@@ -5,7 +5,7 @@ import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import com.mobiquity.mobtravelapp.exception.IncorrectFormatException;
 import com.mobiquity.mobtravelapp.exception.WeatherException;
-import com.mobiquity.mobtravelapp.model.travelModel.*;
+import com.mobiquity.mobtravelapp.model.travel.*;
 import com.mobiquity.mobtravelapp.validation.TravelValidation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
