+++ mobtravelapp/src/test/java/com/mobiquity/mobtravelapp/service/NsServiceTest.java
@@ -1,9 +1,8 @@
 package com.mobiquity.mobtravelapp.service;
 
-import com.mobiquity.mobtravelapp.model.travelModel.RouteModel;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import com.mobiquity.mobtravelapp.model.travel.RouteModel;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentMatchers;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
@@ -12,7 +11,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.util.ReflectionTestUtils;
 import org.springframework.web.client.RestTemplate;
 
@@ -24,7 +22,6 @@ import java.util.stream.Stream;
 
 import static org.junit.Assert.assertNotNull;
 
-@RunWith(SpringRunner.class)
 @SpringBootTest
 public class NsServiceTest {
     @Value("${api.ns.nl.url}")
@@ -39,8 +36,8 @@ public class NsServiceTest {
     @Mock
     RestTemplate restTemplate;
 
-    @Before
-    public void setUp() {
+    @BeforeEach
+    void setUp() {
         ReflectionTestUtils.setField(nsService, "uri", nsUri);
         ReflectionTestUtils.setField(nsService, "key", nsKey);
         Mockito.when(restTemplate.exchange(
