+++ mobtravelapp/src/test/resources/application.properties
@@ -3,3 +3,5 @@ api.darksky.net.url=https://api.darksky.net/forecast/{0}/{1},{2},{3}
 api.ns.nl.key=NSAPIKEY
 api.darksky.net.key=DARKSKYAPIKEY
 api.logger=LoggerFactory.getLogger(TravelService.class)
+
+server.port=80
